Vpd 1812 - Declare vaping products duty new content (#107)
* [VPD-1812] AI assisted - updated form validation

* [VPD-1812] AI assisted - updated type following form refactor

* [VPD-1812] updated content

* [VPD-1812] updated error message

+++ app/controllers/returns/submit/EnterDutyAmountController.scala
@@ -45,7 +45,7 @@ class EnterDutyAmountController @Inject()(
                                         view: EnterDutyAmountView
                                       )(implicit ec: ExecutionContext) extends FrontendBaseController with I18nSupport {
 
-  val form: Form[Int] = formProvider()
+  val form: Form[BigDecimal] = formProvider()
 
   def onPageLoad(mode: Mode): Action[AnyContent] = (identify andThen returnsEnabledAction andThen getData andThen requireData) {
     implicit request =>
